# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Incognito Privacy Platform API
  description: Public REST API for interacting with the Incognito Privacy request engine.
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/v1
    description: Development Server
paths:
...
  /admin/users:
    get:
      summary: List all users
      security:
        - bearerAuth: []
      tags:
        - Admin
      responses:
        '200':
          description: Paginated user index.
  /admin/brokers:
    get:
      summary: List all brokers
      security:
        - bearerAuth: []
      tags:
        - Admin
      responses:
        '200':
          description: All brokers.
    post:
      summary: Create data broker record
      security:
        - bearerAuth: []
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Broker created successfully.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT